Move HotelSearchViewController test helpers to a separated file
